Updated cooldown screen validity check to use uptime rather than just a simple boolean
Earlier, we were just checking to see if the screen ever came on during
cooldown phase. If it did, we would mark the test invalid. However, this
may occur due to the phone receiving notifications, etc which would render
a lot of tests invalid.

This commit introduces a check on the system's uptime. The idea behind
this approach is that the system's uptime does not progress while the
device is asleep. Thus, we do a simple check to make sure that the
device was asleep for at least half of the total cooldown duration
by ensuring that the total amount of elapsed uptime does not exceed
50% of the cooldown period. This 50% threshold may need to be tweaked.