test-device: Replaced vue events with emittery
This has the added benefit of being able to await until all event
handlers are done executing.

Before this, certain async sweep-device event handlers such as the one
which updated the testResult with thermabox data would never have a chance
to complete before the results were shipped off to the server.